requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
Flask==3.1.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
outcome==1.3.0.post0
pandas==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
